+++ modules/katamari/src/test/ts/atomic/api/arr/ObjSizeTest.ts
@@ -15,18 +15,16 @@ describe('atomic.katamari.api.arr.ObjSizeTest', () => {
     check(3, { a: 'a', b: 'b', c: 'c' });
   });
 
-  // Phantomjs gives incorrect results for this test
-  if (navigator.userAgent.indexOf('PhantomJS') <= -1) {
-    // NOTE: This breaks for empty keys
-    it('inductive case', () => {
-      fc.assert(fc.property(
-        fc.dictionary(fc.asciiString(1, 30), fc.integer()),
-        fc.asciiString(1, 30),
-        fc.integer(),
-        (obj, k, v) => {
-          const objWithoutK = Obj.filter(obj, (x, i) => i !== k);
-          assert.deepEqual(Obj.size({ k: v, ...objWithoutK }), Obj.size(objWithoutK) + 1);
-        }));
+  it('inductive case', () => {
+    fc.assert(fc.property(
+      fc.dictionary(fc.asciiString(1, 30), fc.integer()),
+      fc.asciiString(1, 30),
+      fc.integer(),
+      (obj, k, v) => {
+        const objWithoutK = Obj.filter(obj, (x, i) => i !== k);
+        assert.deepEqual(Obj.size({ [k]: v, ...objWithoutK }), Obj.size(objWithoutK) + 1);
+      }), {
+      numRuns: 5000
     });
-  }
+  });
 });