+++ modules/katamari/src/test/ts/atomic/api/fun/ThrottlerTest.ts
@@ -34,7 +34,7 @@ describe('atomic.katamari.api.fun.ThrottlerTest', () => {
       if (value === 'retry') {
         throttler.throttle('retried');
       }
-    }, 100);
+    }, 75);
 
     throttler.throttle('retry');