+++ modules/tinymce/src/core/test/ts/browser/EditorViewTest.ts
@@ -1,4 +1,3 @@
-import { PhantomSkipper } from '@ephox/agar';
 import { before, context, describe, it } from '@ephox/bedrock-client';
 import { Arr } from '@ephox/katamari';
 import { Css, Scroll, SelectorFind } from '@ephox/sugar';
@@ -9,7 +8,6 @@ import Editor from 'tinymce/core/api/Editor';
 import * as EditorView from 'tinymce/core/EditorView';
 
 describe('browser.tinymce.core.EditorViewTest', () => {
-  PhantomSkipper.bddSetup();
 
   const getEditorRect = (editor: Editor) => {
     if (editor.inline) {