+++ modules/tinymce/src/plugins/paste/test/ts/webdriver/CutTest.ts
@@ -1,18 +1,11 @@
-import { PhantomSkipper, RealMouse, Waiter } from '@ephox/agar';
-import { before, describe, it } from '@ephox/bedrock-client';
+import { RealMouse, Waiter } from '@ephox/agar';
+import { describe, it } from '@ephox/bedrock-client';
 import { TinyAssertions, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
 import Plugin from 'tinymce/plugins/paste/Plugin';
 
 describe('webdriver.tinymce.plugins.paste.CutTest', () => {
-  before(function () {
-    // Cut doesn't seem to work in PhantomJS
-    if (PhantomSkipper.detect()) {
-      this.skip();
-    }
-  });
-
   const hook = TinyHooks.bddSetup<Editor>({
     base_url: '/project/tinymce/js/tinymce',
     toolbar: false,