+++ modules/tinymce/src/themes/silver/test/ts/browser/skin/OxideCollectionComponentTest.ts
@@ -1,4 +1,4 @@
-import { ApproxStructure, Assertions, FocusTools, Keys, Mouse, PhantomSkipper, StructAssert, UiFinder } from '@ephox/agar';
+import { ApproxStructure, Assertions, FocusTools, Keys, Mouse, StructAssert, UiFinder } from '@ephox/agar';
 import { TestHelpers } from '@ephox/alloy';
 import { before, context, describe, it } from '@ephox/bedrock-client';
 import { Arr, Optional, Optionals } from '@ephox/katamari';
@@ -136,11 +136,7 @@ describe('browser.tinymce.themes.silver.skin.OxideCollectionComponentTest', () =
       await FocusTools.pTryOnSelector('Focus should be on C', doc, '.tox-collection__item:contains(C)');
     });
 
-    it('Checking the second collection: columns = auto', async function () {
-      // NOTE: We need a layout engine to use flex-wrap navigation.
-      if (PhantomSkipper.detect()) {
-        this.skip();
-      }
+    it('Checking the second collection: columns = auto', async () => {
       const editor = hook.editor();
       const doc = SugarDocument.getDocument();
       FocusTools.setFocus(SugarBody.body(), '.tox-collection__item:contains("C")');
@@ -153,8 +149,7 @@ describe('browser.tinymce.themes.silver.skin.OxideCollectionComponentTest', () =
           children: [
             s.element('div', {
               classes: [ arr.has('tox-collection__group') ],
-              children: Arr.map([ 'D', 'E', 'F' ], (letter) =>
-                structureItem(Optional.none(), Optional.some('icon-' + letter))(s, str, arr)
+              children: Arr.map([ 'D', 'E', 'F' ], (letter) => structureItem(Optional.none(), Optional.some('icon-' + letter))(s, str, arr)
               )
             })
           ]